backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pydantic 
passlib[bcrypt]
python-dotenv